[MISC] move type_traits/template_inspection to detail
