+++ include/seqan3/alignment/matrix/debug_matrix.hpp
@@ -21,7 +21,7 @@
 #include <seqan3/core/detail/debug_stream_range.hpp>
 #include <seqan3/core/detail/debug_stream_type.hpp>
 #include <seqan3/core/detail/debug_stream_optional.hpp>
-#include <seqan3/core/type_traits/template_inspection.hpp>
+#include <seqan3/core/detail/template_inspection.hpp>
 
 namespace seqan3::detail
 {
