+++ include/seqan3/alignment/pairwise/alignment_result.hpp
@@ -17,7 +17,7 @@
 
 #include <seqan3/core/algorithm/configuration.hpp>
 #include <seqan3/core/detail/debug_stream_type.hpp>
-#include <seqan3/core/type_traits/template_inspection.hpp>
+#include <seqan3/core/detail/template_inspection.hpp>
 
 namespace seqan3::detail
 {
