+++ include/seqan3/alignment/pairwise/detail/policy_optimum_tracker.hpp
@@ -20,7 +20,7 @@
 #include <seqan3/alignment/pairwise/detail/type_traits.hpp>
 #include <seqan3/core/algorithm/configuration.hpp>
 #include <seqan3/core/concept/tuple.hpp>
-#include <seqan3/core/type_traits/template_inspection.hpp>
+#include <seqan3/core/detail/template_inspection.hpp>
 
 namespace seqan3::detail
 {
