+++ include/seqan3/core/detail/debug_stream_variant.hpp
@@ -15,7 +15,7 @@
 #include <variant>
 
 #include <seqan3/core/detail/debug_stream_type.hpp>
-#include <seqan3/core/type_traits/template_inspection.hpp>
+#include <seqan3/core/detail/template_inspection.hpp>
 #include <seqan3/utility/type_traits/basic.hpp>
 
 namespace seqan3
