+++ include/seqan3/core/simd/view_to_simd.hpp
@@ -12,20 +12,21 @@
 
 #pragma once
 
+#include <seqan3/std/algorithm>
+#include <seqan3/std/iterator>
+#include <seqan3/std/ranges>
+
 #include <seqan3/alphabet/concept.hpp>
 #include <seqan3/core/detail/pack_algorithm.hpp>
+#include <seqan3/core/detail/template_inspection.hpp>
 #include <seqan3/core/simd/concept.hpp>
 #include <seqan3/core/simd/simd_algorithm.hpp>
 #include <seqan3/core/simd/simd_traits.hpp>
 #include <seqan3/core/simd/simd.hpp>
 #include <seqan3/core/type_traits/range.hpp>
-#include <seqan3/core/type_traits/template_inspection.hpp>
 #include <seqan3/range/views/detail.hpp>
 #include <seqan3/range/views/type_reduce.hpp>
 #include <seqan3/range/views/zip.hpp>
-#include <seqan3/std/algorithm>
-#include <seqan3/std/iterator>
-#include <seqan3/std/ranges>
 
 namespace seqan3::detail
 {
