+++ include/seqan3/io/detail/magic_header.hpp
@@ -20,7 +20,7 @@
 
 #include <seqan3/core/bit_manipulation.hpp>
 #include <seqan3/core/concept/core_language.hpp>
-#include <seqan3/core/type_traits/template_inspection.hpp>
+#include <seqan3/core/detail/template_inspection.hpp>
 #include <seqan3/utility/type_pack/traits.hpp>
 
 namespace seqan3::detail
