+++ include/seqan3/io/record.hpp
@@ -14,7 +14,7 @@
 
 #include <tuple>
 
-#include <seqan3/core/type_traits/template_inspection.hpp>
+#include <seqan3/core/detail/template_inspection.hpp>
 
 namespace seqan3
 {
