+++ include/seqan3/search/search_result.hpp
@@ -16,10 +16,10 @@
 #include <seqan3/std/concepts>
 #include <exception>
 
+#include <seqan3/core/algorithm/configuration.hpp>
 #include <seqan3/core/detail/debug_stream_type.hpp>
 #include <seqan3/core/detail/empty_type.hpp>
-#include <seqan3/core/type_traits/template_inspection.hpp>
-#include <seqan3/core/algorithm/configuration.hpp>
+#include <seqan3/core/detail/template_inspection.hpp>
 #include <seqan3/search/fm_index/concept.hpp>
 
 namespace seqan3::detail
