+++ include/seqan3/utility/type_list/type_list.hpp
@@ -14,7 +14,7 @@
 
 #include <meta/meta.hpp>
 
-#include <seqan3/core/type_traits/template_inspection.hpp>
+#include <seqan3/core/detail/template_inspection.hpp>
 
 namespace seqan3
 {
