+++ test/snippet/core/type_traits/template_inspection_usage.cpp
@@ -1,4 +1,4 @@
-#include <seqan3/core/type_traits/template_inspection.hpp>
+#include <seqan3/core/detail/template_inspection.hpp>
 #include <seqan3/utility/type_list/type_list.hpp>
 
 int main()
