+++ test/snippet/core/type_traits/template_inspection_usage_2.cpp
@@ -1,6 +1,6 @@
 #include <vector>
 
-#include <seqan3/core/type_traits/template_inspection.hpp>
+#include <seqan3/core/detail/template_inspection.hpp>
 #include <seqan3/utility/type_list/type_list.hpp>
 
 int main()
