+++ test/unit/core/type_traits/template_inspection_test.cpp
@@ -9,7 +9,7 @@
 
 #include <gtest/gtest.h>
 
-#include <seqan3/core/type_traits/template_inspection.hpp>
+#include <seqan3/core/detail/template_inspection.hpp>
 #include <seqan3/utility/type_list/type_list.hpp>
 #include <seqan3/utility/type_traits/concept.hpp>
